// src/lib/api/types/api.ts
export interface components {
  schemas: {
// ...
    PermissionResponse: {
      /**
       * @description ID de permiso único
       * @example user:read
       */
      id: string;
      /**
       * @description Nombre del permiso
       * @example user:read
       */
      name: string;
      /**
       * @description Nombre del recurso
       * @example user
       */
      resource: components["schemas"]["ResourceType"];
      /**
       * @description Nombre de la acción
       * @example read
       */
      action: components["schemas"]["ActionType"];
      /**
       * @description Descripción del permiso
       * @example Allows reading user information
       */
      description: string;
      /**
       * Format: date-time
       * @description Fecha de creación del permiso
       * @example 2024-01-01T00:00:00.000Z
       */
      createdAt?: string;
      /**
       * Format: date-time
       * @description Fecha de última actualización del permiso
       * @example 2024-01-01T00:00:00.000Z
       */
      updatedAt?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}